requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.5
python-socketio==5.9.0
python-engineio==4.7.1
PyJWT==2.8.0
python-dotenv==1.0.0
eventlet==0.33.3
redis==4.6.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-socketio==0.0.5
coverage==7.3.2
black==23.9.1
flake8==6.1.0
gunicorn==21.2.0
